계좌마다 수수료를 설정할 수 있다

+++ src/main/scala/com/easywritten/allowancechart/domain/Money.scala
@@ -24,10 +24,9 @@ final case class Money(currency: Currency, amount: MoneyAmount) {
       )
   }
 
-  def *(i: Int): Money = copy(amount = amount * i)
+  def *(i: BigDecimal): Money = copy(amount = amount * i)
 
-  // TODO currency 따라서 scaling
-  def /(i: Int): Money = copy(amount = amount / i)
+  def /(i: BigDecimal): Money = copy(amount = amount / i)
 
 }
